% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart_2.R
\name{pie_chart_2}
\alias{pie_chart_2}
\title{pie_chart_2}
\usage{
pie_chart_2(
  dimension,
  first,
  second = NULL,
  topn = 4,
  titre_1 = "first",
  titre_2 = "second",
  title_yes_no = TRUE,
  dataframe = FALSE
)
}
\arguments{
\item{dimension}{dimension for grouppiing}

\item{first}{first table}

\item{second}{second table}

\item{topn}{number before groupping in Others}

\item{titre_1}{name of first table}

\item{titre_2}{name of second table}

\item{title_yes_no}{is title displayed}

\item{dataframe}{dataframe as image}
}
\value{
Several pie maps
}
\description{
pie_chart_2
}
