% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_knit_child.R
\name{resume_knit_child}
\alias{resume_knit_child}
\title{resume_knit_child}
\usage{
resume_knit_child(dimension, first, second = NULL)
}
\arguments{
\item{dimension}{dimension for groupping}

\item{first}{first table}

\item{second}{second table}
}
\value{
Knit_child of pie and table
}
\description{
resume_knit_child
}
