% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_knit_child.R
\name{resume_knit_child}
\alias{resume_knit_child}
\title{pie_chart_2}
\usage{
resume_knit_child(dimension, first, second = NULL)
}
\arguments{
\item{dimension}{dimension for grouppiing}

\item{first}{first table}

\item{second}{second table}

\item{topn}{number before groupping in Others}

\item{titre_1}{name of first table}

\item{titre_2}{name of second table}
}
\value{
Several pie maps
}
\description{
pie_chart_2
}
